Everything is working, but without tests
I think I've made the explosion stuff a little bit different as expected. When the missile hits an asteroid, the programm creates an Explosion object that will follow the asteroid until it exits the screen or the ammount of frames that it can be on screen ends (which is 20 frames, i.e., 0.3 second). Futhermore, I put a property in each asteroid to allow a displayment with angle of the explosion; this information is sent to the frontend through the field json.

About the boards (scoreboard and leaderboard), they are saved only when
the programm exists when accessing the path /exit (localhost:3000/exit).
This happens because the method I found to make the data saving look nice was
using the java shutdown hook. Apparently, it isn't working when
shutting down the container with ctrl + c.

No tests yet.